chess_game.py
This chess game code is a chess engine with an estimated elo of about 400-700 but this is not accurate it could be more. This code uses pygame and uses a chess engine called NV_Chess_engine.py you need that file in the same directory as this file. You have to change the paths of the images you can use your own but you can find it in assets in this repository.